libjpeg-turbo: Be sure that we should compile in ARM mode
The measurable differences between a binary
compiled in Thumb mode and a binary compiled
in arm mode are very small.

Based on the following benchmarks, it would
seem that arm mode is slightly better:

With 32-bit and Thumb:
Throughput:         70.533766 Megapixels/sec

With 32-bit and arm:
Throughput:         71.082641 Megapixels/sec

The above results were benchmarked using TJBench
(version 1.5.1) using this command:
'tj32 [test-image] 95 -rgb -nowrite -warmup 10'

The above results are the average of 10 runs.

Change-Id: Ibf503c5f01e8fcfc23c3510b1a5ccf368d8a48de
Signed-off-by: Alex Naidis <alex.naidis@linux.com>